feat: 로그인 후 nav bar에 user 정보 보임
